+++ Samples~/Scripts/IssueAndTesting/Issue/Issue1/AssetBundleBuildWindow.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 
 namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue1
@@ -5,5 +6,15 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue.Issue1
     public class AssetBundleBuildWindow : MonoBehaviour
     {
         [SerializeField, Expandable] private AbConfig _abConfig;
+
+        [SerializeField, Expandable] private AbConfig[] _abConfigs;
+
+        [Serializable]
+        public struct ExpandMe
+        {
+            public int intHere;
+        }
+
+        public ExpandMe defaultUnityExpandDraw;
     }
 }